+++ src/main/java/com/packt/app/artist/ArtistRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface ArtistRepository extends CrudRepository<Artist, Integer> {
+public interface ArtistRepository extends CrudRepository<Artist, Long> {
 
 }
